# Introduce Temple generator and freeze string literals