--- a/lib/haml/generator.rb
+++ b/lib/haml/generator.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Haml
   class Generator
     include Temple::Mixins::CompiledDispatcher
@@ -5,6 +6,10 @@ module Haml
 
     def call(exp)
       compile(exp)
+    end
+
+    def on_multi(*exp)
+      exp.map { |e| compile(e) }.join('; ')
     end
 
     def on_code(exp)