--- a/lib/haml/temple_engine.rb
+++ b/lib/haml/temple_engine.rb
@@ -28,6 +28,7 @@ module Haml
 
     use :Parser,   -> { options[:parser_class] }
     use :Compiler, -> { options[:compiler_class] }
+    filter :StaticMerger
     use Generator
 
     def compile(template)