ref: https://github.com/haml/haml/issues/851

I inserted `Haml::Generator` like `Temple::Generator` into `Haml::TempleEngine`. It allows us to use Temple AST to generate compiled Ruby code.

Using that, I added `.freeze` for static string literals for Ruby 2.1+. Whille I'm going to add some changes, this is just a first step.

## Benchmark
With [hamlit/benchmark/etc/tags_loop.haml](https://github.com/k0kubun/hamlit/blob/9a0378ed8a09734655a5813a65378a9c157449b3/benchmark/etc/tags_loop.haml) and Ruby 2.4.0,

### before
```
$ bin/bench benchmark/etc/tags_loop.haml
=================================================
 Rendering: benchmark/etc/tags_loop.haml
=================================================
Calculating -------------------------------------
  haml v5.0.0.beta.2     3.538k i/100ms
         faml v0.8.1     4.801k i/100ms
       hamlit v2.7.5     4.834k i/100ms
-------------------------------------------------
  haml v5.0.0.beta.2     38.358k i/s (0.026ms) -    194.590k
         faml v0.8.1     52.347k i/s (0.019ms) -    259.254k
       hamlit v2.7.5     53.066k i/s (0.019ms) -    265.870k

Comparison:
       hamlit v2.7.5:    53066.3 i/s (0.019ms)
         faml v0.8.1:    52346.5 i/s (0.019ms) - 1.01x slower
  haml v5.0.0.beta.2:    38358.0 i/s (0.026ms) - 1.38x slower
```

### after
```
$ bin/bench benchmark/etc/tags_loop.haml
=================================================
 Rendering: benchmark/etc/tags_loop.haml
=================================================
Calculating -------------------------------------
  haml v5.0.0.beta.2     4.336k i/100ms
         faml v0.8.1     5.150k i/100ms
       hamlit v2.7.5     4.975k i/100ms
-------------------------------------------------
  haml v5.0.0.beta.2     43.109k i/s (0.023ms) -    216.800k
         faml v0.8.1     51.425k i/s (0.019ms) -    257.500k
       hamlit v2.7.5     52.881k i/s (0.019ms) -    263.675k

Comparison:
       hamlit v2.7.5:    52880.7 i/s (0.019ms)
         faml v0.8.1:    51425.4 i/s (0.019ms) - 1.03x slower
  haml v5.0.0.beta.2:    43109.0 i/s (0.023ms) - 1.23x slower
```

Haml's rendering became a little faster.