@amatsuda 2017-02-01
TBH I was thinking of releasing 5.0.0 before such big changes, and taking the next big step at 6.0, but it's done already, and the changes must be good things anyway, so let's do everything @k0kubun wants to do now!